Trait counter::common::web::WebInfo

pub trait WebInfo: Sync + Send {
    type Serialized: Serialize + Send;

    fn get_serializable(&self) -> Self::Serialized;
}

Trait to unify the info that can be send to the web interface.

This has to be implemented for all S in NetworkNodeHandle<S>, e.g. CounterState, ConductorState, etc.

Required Associated Types

type Serialized: Serialize + Send

Required Methods

fn get_serializable(&self) -> Self::Serialized