Struct counter::Counter

pub struct Counter<T: Hash + Eq, N = usize> { /* private fields */ }

Implementations

impl<T, N> Counter<T, N>where T: Hash + Eq,

pub fn into_map(self) -> HashMap<T, N>

Consumes this counter and returns a HashMap mapping the items to the counts.

pub fn total<'a, S>(&'a self) -> Swhere S: Sum<&'a N>,

Returns the sum of the counts.

Use len to get the number of elements in the counter and use total to get the sum of their counts.

Examples

let counter = Counter::init("abracadabra".chars());
assert_eq!(counter.total::<usize>(), 11);
assert_eq!(counter.len(), 5);

impl<T, N> Counter<T, N>where T: Hash + Eq, N: Zero,

pub fn new() -> Counter<T, N>

Create a new, empty Counter

impl<T, N> Counter<T, N>where T: Hash + Eq, N: AddAssign + Zero + One,

pub fn init<I>(iterable: I) -> Counter<T, N>where I: IntoIterator<Item = T>,

Create a new Counter initialized with the given iterable.

pub fn update<I>(&mut self, iterable: I)where I: IntoIterator<Item = T>,

Add the counts of the elements from the given iterable to this counter.

impl<T, N> Counter<T, N>where T: Hash + Eq, N: PartialOrd + SubAssign + Zero + One,

pub fn subtract<I>(&mut self, iterable: I)where I: IntoIterator<Item = T>,

Remove the counts of the elements from the given iterable to this counter.

Non-positive counts are automatically removed.

let mut counter = "abbccc".chars().collect::<Counter<_>>();
counter.subtract("abba".chars());
let expect = [('c', 3)].iter().cloned().collect::<HashMap<_, _>>();
assert_eq!(counter.into_map(), expect);

impl<T, N> Counter<T, N>where T: Hash + Eq + Clone, N: Clone + Ord,

pub fn most_common(&self) -> Vec<(T, N)>

Create a vector of (elem, frequency) pairs, sorted most to least common.

let mc = "pappaopolo".chars().collect::<Counter<_>>().most_common();
let expected = vec![('p', 4), ('o', 3), ('a', 2), ('l', 1)];
assert_eq!(mc, expected);

Note that the ordering of duplicates is unstable.

pub fn most_common_tiebreaker<F>(&self, tiebreaker: F) -> Vec<(T, N)>where F: FnMut(&T, &T) -> Ordering,

Create a vector of (elem, frequency) pairs, sorted most to least common.

In the event that two keys have an equal frequency, use the supplied ordering function to further arrange the results.

For example, we can sort reverse-alphabetically:

let counter = "eaddbbccc".chars().collect::<Counter<_>>();
let by_common = counter.most_common_tiebreaker(|&a, &b| b.cmp(&a));
let expected = vec![('c', 3), ('d', 2), ('b', 2), ('e', 1), ('a', 1)];
assert_eq!(by_common, expected);

impl<T, N> Counter<T, N>where T: Hash + Eq + Clone + Ord, N: Clone + Ord,

pub fn most_common_ordered(&self) -> Vec<(T, N)>

Create a vector of (elem, frequency) pairs, sorted most to least common.

In the event that two keys have an equal frequency, use the natural ordering of the keys to further sort the results.

Examples

let mc = "abracadabra".chars().collect::<Counter<_>>().most_common_ordered();
let expect = vec![('a', 5), ('b', 2), ('r', 2), ('c', 1), ('d', 1)];
assert_eq!(mc, expect);

Time complexity

O(n * log n), where n is the number of items in the counter. If all you want is the top k items and k < n then it can be more efficient to use k_most_common_ordered.

pub fn k_most_common_ordered(&self, k: usize) -> Vec<(T, N)>

Returns the k most common items in decreasing order of their counts.

The returned vector is the same as would be obtained by calling most_common_ordered and then truncating the result to length k. In particular, items with the same count are sorted in increasing order of their keys. Further, if k is greater than the length of the counter then the returned vector will have length equal to that of the counter, not k.

Examples

let counter: Counter<_> = "abracadabra".chars().collect();
let top3 = counter.k_most_common_ordered(3);
assert_eq!(top3, vec![('a', 5), ('b', 2), ('r', 2)]);

Time complexity

This method can be much more efficient than most_common_ordered when k is much smaller than the length of the counter n. When k = 1 the algorithm is equivalent to finding the minimum (or maximum) of n items, which requires n - 1 comparisons. For a fixed value of k > 1, the number of comparisons scales with n as n + O(log n) and the number of swaps scales as O(log n). As k approaches n, this algorithm approaches a heapsort of the n items, which has complexity O(n * log n).

For values of k close to n the sorting algorithm used by most_common_ordered will generally be faster than the heapsort used by this method by a small constant factor. Exactly where the crossover point occurs will depend on several factors. For small k choose this method. If k is a substantial fraction of n, it may be that most_common_ordered is faster. If performance matters in your application then it may be worth experimenting to see which of the two methods is faster.

impl<T, N> Counter<T, N>where T: Hash + Eq, N: PartialOrd + Zero,

pub fn is_superset(&self, other: &Self) -> bool

Test whether this counter is a superset of another counter. This is true if for all elements in this counter and the other, the count in this counter is greater than or equal to the count in the other.

c.is_superset(&d); -> c.iter().all(|(x, n)| n >= d[x]) && d.iter().all(|(x, n)| c[x] >= n)

let c = "aaabbc".chars().collect::<Counter<_>>();
let mut d = "abb".chars().collect::<Counter<_>>();

assert!(c.is_superset(&d));
d[&'e'] = 1;
assert!(!c.is_superset(&d));

pub fn is_subset(&self, other: &Self) -> bool

Test whether this counter is a subset of another counter. This is true if for all elements in this counter and the other, the count in this counter is less than or equal to the count in the other.

c.is_subset(&d); -> c.iter().all(|(x, n)| n <= d[x]) && d.iter().all(|(x, n)| c[x] <= n)

let mut c = "abb".chars().collect::<Counter<_>>();
let mut d = "aaabbc".chars().collect::<Counter<_>>();

assert!(c.is_subset(&d));
c[&'e'] = 1;
assert!(!c.is_subset(&d));